# openapi.yaml
openapi: 3.1.0
info:
  title: MatchTracker Pro API
  version: 0.1.0
  description: >
    Contract between the MatchTracker Pro frontend (currently mocked) and the
    FastAPI backend. Phase 1 backend may use an in-memory store; Phase 2 moves
    to SQLAlchemy + SQLite without changing this contract.
servers:
  - url: /api

paths:
  /teams:
    get:
      operationId: listTeams
      summary: List available teams to follow
      responses:
        "200":
          description: Teams
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Team" }

  /subscriptions:
    get:
      operationId: getSubscription
      summary: Current user's subscription preferences
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Subscription
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Subscription" }
        "401":
          description: Unauthorized
    post:
      operationId: saveSubscription
      summary: Save selected teams and reminder intervals
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubscriptionPayload" }
      responses:
        "200":
          description: Updated subscription
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Subscription" }
        "401":
          description: Unauthorized

  /matches/urgency:
    get:
      operationId: listMatchesByUrgency
      summary: Upcoming subscribed matches sorted by time to kickoff
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Matches
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Match" }
        "401":
          description: Unauthorized

  /matches/{id}/bet-placed:
    post:
      operationId: markBetPlaced
      summary: Toggle "Bet Placed" status, snoozing betting reminders
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Updated match
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Match" }
        "401":
          description: Unauthorized
        "404":
          description: Match not found

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token for user authentication
  schemas:
    Team:
      type: object
      required: [id, name, league, crestUrl]
      properties:
        id: { type: string }
        name: { type: string }
        league: { type: string }
        crestUrl: { type: string }

    Match:
      type: object
      required:
        [id, homeTeam, awayTeam, competition, venue, kickoffAt, crestUrl, urgency, betPlaced]
      properties:
        id: { type: string }
        homeTeam: { type: string }
        awayTeam: { type: string }
        competition: { type: string }
        venue: { type: string }
        kickoffAt: { type: string, format: date-time }
        crestUrl: { type: string }
        urgency:
          type: string
          enum: [critical, soon, later]
        betPlaced: { type: boolean }

    ReminderInterval:
      type: string
      enum: ["5d", "2d", "2h"]

    SubscriptionPayload:
      type: object
      required: [teamIds, reminderIntervals]
      properties:
        teamIds:
          type: array
          items: { type: string }
        reminderIntervals:
          type: array
          items: { $ref: "#/components/schemas/ReminderInterval" }

    Subscription:
      allOf:
        - $ref: "#/components/schemas/SubscriptionPayload"
        - type: object
          required: [id, updatedAt]
          properties:
            id: { type: string }
            telegramHandle:
              type: [string, "null"]
            updatedAt: { type: string, format: date-time }